Add a guide document and update thumbnails
Update blender file so that thumbnails are rendered using lineart rather
than freestyle. This required some tweaks to ensure that sharp edges
weren't set overzealously on the mantle surface and edges of plates.
Lighting is added, and background changed to white. It also utilises a
version of the "Screen Space Thickness" node group [1] modified to work
with orthographic cameras.

Also add a guide document, using graphics rendered in a similar way, but
in separate scenes, and update the .gitignore.

The render script has an import_all variable added to allow obj files to
be easily imported for testing, and a material_emit_white variable to
allow imported objects to be tweaked to appear white (useful for printer
friendliness of the guide).

The render script is also updated to handle the "3.3. Guide - Plates"
scene, which can have multiple models loaded per camera (for when a
plate is sliced), and different resolutions to render at in order to
achieve a desired DPI for printing (important to get the lineart looking
right).

A new "Load objs into selected empties" script is used for the "3.1.
Guide - Introduction" and "3.2. Guide - Building Interior" scenes. It
loads objects and parents them under empties to allow them to be
animated. The rendered animations goes into `guide/` where the frames
are used by the guide SVG. The objects are loaded into a collection,
which should be undone or the collection cleared out after use.

Signed-off-by: James Hogan <james@albanarts.com>

[1] https://developer.blender.org/docs/release_notes/4.3/grease_pencil_migration/#screen-space-stroke-thickness