Adjust voxel size to 0.2mm
The overall size of the final 3mf is too large to upload to MakerWorld.
The remesh voxel size in the final modifier is 0.15mm, which is finer
than the 0.2mm layer size the model is printed with. Lets try and reduce
the final size of the meshes by increasing the voxel size to 0.2mm. It
is also now driven by a "voxel size" custom property of the "Model
Configuration" object.

Signed-off-by: James Hogan <james@albanarts.com>