Double magnets between mantle segments (#1)
Improve adhesion between mantle segments by doubling the number of
magnets. Each magnet is moved 25 degrees from the previous central
position, and the "Mantle" labels are centered to where the magnets
were. The northern hemisphere mantle segment identifiers are also moved
to be near the north pole, for consistency with the southern hemisphere
segments.

In order to render lineart similarly with Blender 4.5, the lineart
thickness has been doubled in the geometry shader.

Signed-off-by: James Hogan <james@albanarts.com>