Plate mesh generation fixes
The okhotsk plate mesh is being generated without a magnet slot due to
broken boolean operations. Fix this by switching the mesh boolean
node algorithm for subtracting magnets in the "Cut plate boundaries and
magnets" geometry nodes from Float to Manifold.

Unfortunately this triggers issues in the boolean slicing of some of the
larger plates, with both pacific_2 and north_america_2 showing up empty.
Fix this similarly, switching the boolean modifier algorithms of all
plate slices to Manifold (with materials transferred). Previously a mix
of Float, Exact, and Manifold (with indexed and transferred materials)
algorithms were used as required to avoid boolean operation failures.

Finally drive the voxel size of the pre-boolean remesh of plate slice
volumes from the model configuration (0.15mm -> 0.2mm) so the slice
edges triangle density matches that of the plate surfaces.

Signed-off-by: James Hogan <james@albanarts.com>